Code.py
Running this code opens a new window to access add, delete and mark as completed functions. When you click on add task a new window opens where you enter the task and again click on add task the task will we added to the list. When you select a task and then click on delete or mark as completed, the task will be deleted or marked as completed resp.